Rename types tests to types
